@@ -1,6 +1,5 @@
 import React from 'react';
-
-const credits =(props) => {
+const credits = (props) => {
     return (
        <div>
            Created by <a href="https://github.com/alond2010">alond2010</a>
